refactor: delete the HyperKZG commitment scheme
HyperKZG had no production caller. The modular `jolt-hyperkzg` crate was
reachable only as a `jolt-openings` dev-dependency, and the legacy
`poly/commitment/hyperkzg.rs` carried its own note that "HyperKZG is not
currently used in Jolt". Dory is the only PCS the prover and verifier
actually instantiate.

Removed:
- crates/jolt-hyperkzg/ — the crate, its criterion bench, and its three
  cargo-fuzz targets (none of which CI ever ran: neither fuzz-crates.yml
  nor bench-crates.yml lists the crate)
- crates/jolt-prover-legacy/src/poly/commitment/{hyperkzg,kzg}.rs — the
  legacy port and the UnivariateKZG/SRS primitives it was the sole
  consumer of
- crates/jolt-openings/tests/homomorphic_hyperkzg.rs

`tests/support/common.rs` is `#[path]`-included by the surviving Dory and
schemes test binaries, so its `kzg_setup` helper and imports go with it.
Its clear-path witness-count-mismatch case covered generic
`HomomorphicBatch` code that survives — Dory only had the ZK-path
equivalent — so that test is ported to homomorphic_dory.rs rather than
dropped.

Docs are updated where they described the workspace as it was: the crate
count, the legacy `poly/` module map, and the verifier-closure crate
count. specs/verifier-closure-lints.md keeps its dated 2026-07-26 survey
verbatim and gains an addendum instead, since it records a campaign as
executed rather than current state.
