--- crates/jolt-hyperkzg/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "jolt-hyperkzg"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-description = "HyperKZG multilinear polynomial commitment scheme for the Jolt zkVM"
-
-[lints]
-workspace = true
-
-[dependencies]
-jolt-crypto = { path = "../jolt-crypto" }
-jolt-field = { path = "../jolt-field" }
-jolt-poly = { path = "../jolt-poly" }
-jolt-transcript = { path = "../jolt-transcript" }
-jolt-openings = { path = "../jolt-openings" }
-serde = { workspace = true, features = ["derive"] }
-tracing.workspace = true
-num-traits = { workspace = true }
-rayon = { workspace = true }
-thiserror = { workspace = true }
-rand_core = { workspace = true, features = ["getrandom"] }
-rand_chacha = { workspace = true }
-
-[dev-dependencies]
-jolt-field = { path = "../jolt-field", features = ["bn254"] }
-criterion = { workspace = true }
-rand = { workspace = true }
-
-[[bench]]
-name = "hyperkzg"
-harness = false
-
-[package.metadata.cargo-machete]
-ignored = ["rand_core", "rand_chacha", "rand"]
