--- crates/jolt-hyperkzg/src/scheme.rs
@@ -1,690 +1,0 @@
-//! HyperKZG commitment scheme implementing `jolt-openings` traits.
-//!
-//! [`HyperKZGScheme`] is generic over `P: PairingGroup` — instantiate with
-//! `Bn254` for the concrete BN254 curve.
-
-use std::marker::PhantomData;
-
-use jolt_crypto::{Commitment, DeriveSetup, JoltGroup, PairingGroup, PedersenSetup};
-use jolt_field::{Field, Ring};
-use jolt_openings::{AdditivelyHomomorphic, CommitmentScheme, OpeningsError};
-use jolt_poly::MultilinearPoly;
-use jolt_transcript::{AppendToTranscript, Transcript};
-use num_traits::{One, Zero};
-use rayon::prelude::*;
-use serde::{de::DeserializeOwned, Serialize};
-
-use crate::error::HyperKZGError;
-use crate::kzg::{self, kzg_open_batch, kzg_verify_batch};
-use crate::types::{HyperKZGCommitment, HyperKZGProof, HyperKZGProverSetup, HyperKZGVerifierSetup};
-
-/// HyperKZG multilinear polynomial commitment scheme.
-///
-/// Generic over `P: PairingGroup`. Implements [`CommitmentScheme`] and
-/// [`AdditivelyHomomorphic`] from `jolt-openings`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct HyperKZGScheme<P: PairingGroup> {
-    _phantom: PhantomData<P>,
-}
-
-impl<P: PairingGroup> HyperKZGScheme<P>
-where
-    P::ScalarField: AppendToTranscript,
-    P::G1: AppendToTranscript,
-{
-    /// Generates an SRS from a random generator and secret scalar.
-    ///
-    /// `max_degree` is the maximum polynomial length (number of evaluations).
-    /// The SRS will contain `max_degree + 1` G1 powers and 2 G2 powers.
-    pub fn setup<R: rand_core::RngCore>(
-        rng: &mut R,
-        max_degree: usize,
-        g1: P::G1,
-        g2: P::G2,
-    ) -> HyperKZGProverSetup<P> {
-        let beta = P::ScalarField::random(rng);
-        Self::setup_from_secret(beta, max_degree, g1, g2)
-    }
-
-    /// Generates SRS from a known secret.
-    ///
-    /// WARNING: this is only appropriate for deterministic tests or trusted
-    /// setup tooling that destroys `beta`; anyone who knows `beta` can break
-    /// KZG binding.
-    pub fn setup_from_secret(
-        beta: P::ScalarField,
-        max_degree: usize,
-        g1: P::G1,
-        g2: P::G2,
-    ) -> HyperKZGProverSetup<P> {
-        let mut g1_powers = Vec::with_capacity(max_degree + 1);
-        let mut cur = g1;
-        for _ in 0..=max_degree {
-            g1_powers.push(cur);
-            cur = cur.scalar_mul(&beta);
-        }
-
-        let g2_powers = vec![g2, g2.scalar_mul(&beta)];
-
-        HyperKZGProverSetup {
-            g1_powers,
-            g2_powers,
-        }
-    }
-
-    /// Phase 1 of the HyperKZG protocol: fold the multilinear polynomial.
-    ///
-    /// Given polynomial $P$ with $2^\ell$ evaluations and opening point
-    /// $x = (x_1, \ldots, x_\ell)$, produces $\ell$ polynomials
-    /// $P_0 = P, P_1, \ldots, P_{\ell-1}$ where each $P_i$ has half
-    /// the length of $P_{i-1}$.
-    ///
-    /// The folding relation is:
-    /// $P_i[j] = (1 - x_{\ell-i}) \cdot P_{i-1}[2j] + x_{\ell-i} \cdot P_{i-1}[2j+1]$
-    #[expect(
-        clippy::expect_used,
-        reason = "polys is seeded with one element before the fold loop"
-    )]
-    fn fold_polynomials(
-        evals: &[P::ScalarField],
-        point: &[P::ScalarField],
-    ) -> Vec<Vec<P::ScalarField>> {
-        let ell = point.len();
-        let mut polys = Vec::with_capacity(ell);
-        polys.push(evals.to_vec());
-
-        // Fold i uses x_{ell-i}, i.e. point[1..] visited back-to-front.
-        for &xi in point.iter().skip(1).rev() {
-            let prev = polys.last().expect("polys starts with one element");
-            let pi: Vec<P::ScalarField> = prev
-                .par_chunks_exact(2)
-                .map(|pair| {
-                    #[expect(
-                        clippy::indexing_slicing,
-                        reason = "par_chunks_exact(2) yields exactly-2-element slices"
-                    )]
-                    let (even, odd) = (pair[0], pair[1]);
-                    even + xi * (odd - even)
-                })
-                .collect();
-            polys.push(pi);
-        }
-
-        polys
-    }
-
-    /// Full HyperKZG opening proof.
-    #[tracing::instrument(skip_all, name = "HyperKZG::open")]
-    #[expect(
-        clippy::expect_used,
-        reason = "intermediate fold polynomials shrink, so an SRS sized for the input covers every kzg_commit"
-    )]
-    pub fn open<T: Transcript<Challenge = P::ScalarField>>(
-        setup: &HyperKZGProverSetup<P>,
-        evals: &[P::ScalarField],
-        point: &[P::ScalarField],
-        transcript: &mut T,
-    ) -> Result<HyperKZGProof<P>, HyperKZGError> {
-        let ell = point.len();
-        if ell == 0 {
-            return Err(HyperKZGError::EmptyPoint);
-        }
-        let n = evals.len();
-        assert_eq!(n, 1 << ell, "evaluation count must be 2^ell");
-
-        // Phase 1: fold
-        let polys = Self::fold_polynomials(evals, point);
-        assert_eq!(polys.len(), ell);
-        assert_eq!(polys.last().map(Vec::len), Some(2));
-
-        // Commit to intermediate polynomials (skip polys[0] — already committed)
-        let com: Vec<P::G1> = polys
-            .par_iter()
-            .skip(1)
-            .map(|p| kzg::kzg_commit::<P>(p, setup).expect("SRS large enough for intermediate"))
-            .collect();
-
-        // Phase 2: derive challenge r
-        for c in &com {
-            transcript.append(c);
-        }
-        let r: P::ScalarField = transcript.challenge();
-        let u = [r, -r, r * r];
-
-        // Phase 3: batch open all polynomials at the three points
-        let (w, v) = kzg_open_batch::<P, T>(&polys, &u, setup, transcript);
-
-        Ok(HyperKZGProof { com, w, v })
-    }
-
-    /// HyperKZG verification.
-    #[tracing::instrument(skip_all, name = "HyperKZG::verify")]
-    pub fn verify<T: Transcript<Challenge = P::ScalarField>>(
-        vk: &HyperKZGVerifierSetup<P>,
-        commitment: &HyperKZGCommitment<P>,
-        point: &[P::ScalarField],
-        claimed_eval: &P::ScalarField,
-        proof: &HyperKZGProof<P>,
-        transcript: &mut T,
-    ) -> Result<(), HyperKZGError> {
-        let ell = point.len();
-        if ell == 0 {
-            return Err(HyperKZGError::EmptyPoint);
-        }
-
-        if proof.com.len() + 1 != ell {
-            return Err(HyperKZGError::WrongCommitmentCount {
-                expected: ell - 1,
-                got: proof.com.len(),
-            });
-        }
-
-        // Validate inner evaluation widths before mutating the transcript.
-        let v = &proof.v;
-        if v[0].len() != ell || v[1].len() != ell || v[2].len() != ell {
-            return Err(HyperKZGError::WrongEvaluationWidth { expected: ell });
-        }
-
-        // Absorb intermediate commitments
-        for c in &proof.com {
-            transcript.append(c);
-        }
-        let r: P::ScalarField = transcript.challenge();
-
-        if r.is_zero() {
-            return Err(HyperKZGError::DegenerateChallenge);
-        }
-
-        // Prepend the original commitment as C_0
-        let mut com = Vec::with_capacity(ell);
-        com.push(commitment.point);
-        com.extend_from_slice(&proof.com);
-
-        let u = [r, -r, r * r];
-
-        let ypos = &v[0]; // evaluations at r
-        let yneg = &v[1]; // evaluations at -r
-        let mut y_sq = v[2].clone(); // evaluations at r^2
-        y_sq.push(*claimed_eval);
-
-        // Consistency check: the folding relation must hold across evaluations
-        //
-        // For each level i, the polynomial P_i is defined by:
-        //   P_i(x) = (1 - x_{ell-i}) * P_{i-1,even}(x) + x_{ell-i} * P_{i-1,odd}(x)
-        //
-        // This implies:
-        //   2*r * P_{i+1}(r^2) = r * (1 - x_{ell-i-1}) * (P_i(r) + P_i(-r))
-        //                       + x_{ell-i-1} * (P_i(r) - P_i(-r))
-        // All four iterators have exactly `ell` elements: the widths were
-        // validated above and `y_sq` carries the extra `claimed_eval` entry.
-        let two = P::ScalarField::from_u64(2);
-        for (level, (((&y_next, &y_pos), &y_neg), &x)) in y_sq
-            .iter()
-            .skip(1)
-            .zip(ypos.iter())
-            .zip(yneg.iter())
-            .zip(point.iter().rev())
-            .enumerate()
-        {
-            let lhs = two * r * y_next;
-            let rhs = r * (P::ScalarField::one() - x) * (y_pos + y_neg) + x * (y_pos - y_neg);
-            if lhs != rhs {
-                return Err(HyperKZGError::FoldingConsistencyFailed { level });
-            }
-        }
-
-        // Batch KZG pairing check
-        if !kzg_verify_batch::<P, T>(vk, &com, &proof.w, &u, &proof.v, transcript) {
-            return Err(HyperKZGError::PairingCheckFailed);
-        }
-
-        Ok(())
-    }
-}
-
-/// # Security note
-///
-/// Uses KZG SRS powers as Pedersen generators — Pedersen binding shares the
-/// KZG trapdoor `beta`. Both are sound once `beta` is destroyed, but the two
-/// schemes do not have independent security assumptions.
-impl<P: PairingGroup> DeriveSetup<HyperKZGProverSetup<P>> for PedersenSetup<P::G1> {
-    /// # Panics
-    ///
-    /// Panics when the SRS is smaller than `capacity + 1`. `derive` runs at
-    /// setup time on operator-provided parameters (the infallible
-    /// `DeriveSetup` trait offers no error channel), never on the
-    /// proof-verification path.
-    #[expect(clippy::expect_used, reason = "length checked by the assert above")]
-    fn derive(source: &HyperKZGProverSetup<P>, capacity: usize) -> Self {
-        assert!(
-            source.g1_powers.len() > capacity,
-            "SRS has {} G1 powers, need at least {} (capacity + 1 for blinding)",
-            source.g1_powers.len(),
-            capacity + 1,
-        );
-        let (message_generators, rest) = source.g1_powers.split_at(capacity);
-        let blinding_generator = *rest.first().expect("length checked by the assert above");
-        PedersenSetup::new(message_generators.to_vec(), blinding_generator)
-    }
-}
-
-impl<P: PairingGroup> Commitment for HyperKZGScheme<P> {
-    type Output = HyperKZGCommitment<P>;
-}
-
-impl<P: PairingGroup> CommitmentScheme for HyperKZGScheme<P>
-where
-    P::ScalarField: AppendToTranscript + Serialize + DeserializeOwned,
-    P::G1: AppendToTranscript + Serialize + DeserializeOwned,
-    P::G2: Serialize + DeserializeOwned,
-{
-    type Field = P::ScalarField;
-    type Proof = HyperKZGProof<P>;
-    type ProverSetup = HyperKZGProverSetup<P>;
-    type VerifierSetup = HyperKZGVerifierSetup<P>;
-    type OpeningHint = ();
-    type SetupParams = (usize, P::G1, P::G2);
-
-    fn setup(
-        (max_num_vars, g1, g2): Self::SetupParams,
-    ) -> Result<(Self::ProverSetup, Self::VerifierSetup), OpeningsError> {
-        let mut rng = rand_core::OsRng;
-        let max_degree = 1usize << max_num_vars;
-        let prover = HyperKZGScheme::setup(&mut rng, max_degree, g1, g2);
-        let verifier = Self::verifier_setup(&prover);
-        Ok((prover, verifier))
-    }
-
-    fn verifier_setup(prover_setup: &Self::ProverSetup) -> Self::VerifierSetup {
-        HyperKZGVerifierSetup::from(prover_setup)
-    }
-
-    fn commit<S: MultilinearPoly<Self::Field> + ?Sized>(
-        poly: &S,
-        setup: &Self::ProverSetup,
-    ) -> Result<(Self::Output, Self::OpeningHint), OpeningsError> {
-        // HyperKZG always works on dense evaluations; `to_dense` borrows them
-        // when the source is already dense.
-        let evaluations = poly.to_dense();
-        let point = kzg::kzg_commit::<P>(&evaluations, setup)
-            .map_err(|e| OpeningsError::CommitFailed(format!("HyperKZG commit failed: {e:?}")))?;
-        Ok((HyperKZGCommitment { point }, ()))
-    }
-
-    fn open<S: MultilinearPoly<Self::Field> + ?Sized>(
-        poly: &S,
-        point: &[Self::Field],
-        _eval: Self::Field,
-        setup: &Self::ProverSetup,
-        _hint: Option<Self::OpeningHint>,
-        transcript: &mut impl Transcript<Challenge = Self::Field>,
-    ) -> Result<Self::Proof, OpeningsError> {
-        let evaluations = poly.to_dense();
-        Self::open(setup, &evaluations, point, transcript)
-            .map_err(|e| OpeningsError::ProveFailed(format!("HyperKZG open failed: {e:?}")))
-    }
-
-    fn verify(
-        commitment: &Self::Output,
-        point: &[Self::Field],
-        eval: Self::Field,
-        proof: &Self::Proof,
-        setup: &Self::VerifierSetup,
-        transcript: &mut impl Transcript<Challenge = Self::Field>,
-    ) -> Result<(), OpeningsError> {
-        Self::verify(setup, commitment, point, &eval, proof, transcript)
-            .map_err(|_| OpeningsError::VerificationFailed)
-    }
-}
-
-impl<P: PairingGroup> AdditivelyHomomorphic for HyperKZGScheme<P>
-where
-    P::ScalarField: AppendToTranscript + Serialize + DeserializeOwned,
-    P::G1: AppendToTranscript + Serialize + DeserializeOwned,
-    P::G2: Serialize + DeserializeOwned,
-{
-    fn combine(commitments: &[Self::Output], scalars: &[Self::Field]) -> Self::Output {
-        assert_eq!(commitments.len(), scalars.len());
-        let bases: Vec<P::G1> = commitments.iter().map(|c| c.point).collect();
-        HyperKZGCommitment {
-            point: P::G1::msm(&bases, scalars),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    #![expect(
-        clippy::unwrap_used,
-        clippy::expect_used,
-        reason = "tests unwrap successful PCS operations"
-    )]
-
-    use super::*;
-    use jolt_crypto::Bn254;
-    use jolt_field::Fr;
-    use jolt_poly::Polynomial;
-    use jolt_transcript::Blake2bTranscript;
-    use rand_chacha::ChaCha20Rng;
-    use rand_core::SeedableRng;
-
-    type TestScheme = HyperKZGScheme<Bn254>;
-
-    fn test_setup(max_degree: usize) -> (HyperKZGProverSetup<Bn254>, HyperKZGVerifierSetup<Bn254>) {
-        let mut rng = ChaCha20Rng::seed_from_u64(0xdead_beef);
-        let g1 = Bn254::g1_generator();
-        let g2 = Bn254::g2_generator();
-        let prover = TestScheme::setup(&mut rng, max_degree, g1, g2);
-        let verifier = TestScheme::verifier_setup(&prover);
-        (prover, verifier)
-    }
-
-    #[test]
-    fn commit_open_verify_roundtrip() {
-        for ell in [2, 3, 4, 6, 8] {
-            let n = 1 << ell;
-            let mut rng = ChaCha20Rng::seed_from_u64(ell as u64);
-            let (pk, vk) = test_setup(n);
-
-            let poly = Polynomial::<Fr>::random(ell, &mut rng);
-            let point: Vec<Fr> = (0..ell).map(|_| Fr::random(&mut rng)).collect();
-            let eval = poly.evaluate(&point);
-
-            let (commitment, ()) = TestScheme::commit(poly.evaluations(), &pk).unwrap();
-
-            let mut prover_transcript = Blake2bTranscript::new(b"test");
-            let proof = <TestScheme as CommitmentScheme>::open(
-                &poly,
-                &point,
-                eval,
-                &pk,
-                None,
-                &mut prover_transcript,
-            )
-            .unwrap();
-
-            let mut verifier_transcript = Blake2bTranscript::new(b"test");
-            let result = <TestScheme as CommitmentScheme>::verify(
-                &commitment,
-                &point,
-                eval,
-                &proof,
-                &vk,
-                &mut verifier_transcript,
-            );
-            assert!(result.is_ok(), "ell={ell}: verification failed: {result:?}");
-        }
-    }
-
-    #[test]
-    fn wrong_eval_rejects() {
-        let ell = 4;
-        let n = 1 << ell;
-        let mut rng = ChaCha20Rng::seed_from_u64(42);
-        let (pk, vk) = test_setup(n);
-
-        let poly = Polynomial::<Fr>::random(ell, &mut rng);
-        let point: Vec<Fr> = (0..ell).map(|_| Fr::random(&mut rng)).collect();
-        let eval = poly.evaluate(&point);
-        let wrong_eval = eval + Fr::from_u64(1);
-
-        let (commitment, ()) = TestScheme::commit(poly.evaluations(), &pk).unwrap();
-
-        let mut prover_transcript = Blake2bTranscript::new(b"test-bad");
-        let proof = <TestScheme as CommitmentScheme>::open(
-            &poly,
-            &point,
-            eval,
-            &pk,
-            None,
-            &mut prover_transcript,
-        )
-        .unwrap();
-
-        let mut verifier_transcript = Blake2bTranscript::new(b"test-bad");
-        let result = <TestScheme as CommitmentScheme>::verify(
-            &commitment,
-            &point,
-            wrong_eval,
-            &proof,
-            &vk,
-            &mut verifier_transcript,
-        );
-        assert!(result.is_err(), "wrong evaluation should be rejected");
-    }
-
-    #[test]
-    fn missing_intermediate_commitment_rejects() {
-        let ell = 4;
-        let n = 1 << ell;
-        let mut rng = ChaCha20Rng::seed_from_u64(43);
-        let (pk, vk) = test_setup(n);
-
-        let poly = Polynomial::<Fr>::random(ell, &mut rng);
-        let point: Vec<Fr> = (0..ell).map(|_| Fr::random(&mut rng)).collect();
-        let eval = poly.evaluate(&point);
-
-        let (commitment, ()) = TestScheme::commit(poly.evaluations(), &pk).unwrap();
-
-        let mut prover_transcript = Blake2bTranscript::new(b"test-missing-com");
-        let mut proof = <TestScheme as CommitmentScheme>::open(
-            &poly,
-            &point,
-            eval,
-            &pk,
-            None,
-            &mut prover_transcript,
-        )
-        .unwrap();
-        let _ = proof.com.pop();
-
-        let mut verifier_transcript = Blake2bTranscript::new(b"test-missing-com");
-        let result = TestScheme::verify(
-            &vk,
-            &commitment,
-            &point,
-            &eval,
-            &proof,
-            &mut verifier_transcript,
-        );
-        assert!(matches!(
-            result,
-            Err(HyperKZGError::WrongCommitmentCount { .. })
-        ));
-    }
-
-    #[test]
-    fn trait_setup_uses_fresh_randomness() {
-        let g1 = Bn254::g1_generator();
-        let g2 = Bn254::g2_generator();
-
-        let (_pk1, vk1) = <TestScheme as CommitmentScheme>::setup((4, g1, g2)).unwrap();
-        let (_pk2, vk2) = <TestScheme as CommitmentScheme>::setup((4, g1, g2)).unwrap();
-
-        assert_ne!(vk1.beta_g2, vk2.beta_g2);
-    }
-
-    #[test]
-    fn tampered_proof_rejects() {
-        let ell = 4;
-        let n = 1 << ell;
-        let mut rng = ChaCha20Rng::seed_from_u64(99);
-        let (pk, vk) = test_setup(n);
-
-        let poly = Polynomial::<Fr>::random(ell, &mut rng);
-        let point: Vec<Fr> = (0..ell).map(|_| Fr::random(&mut rng)).collect();
-        let eval = poly.evaluate(&point);
-
-        let (commitment, ()) = TestScheme::commit(poly.evaluations(), &pk).unwrap();
-
-        let mut prover_transcript = Blake2bTranscript::new(b"test-tamper");
-        let mut proof = <TestScheme as CommitmentScheme>::open(
-            &poly,
-            &point,
-            eval,
-            &pk,
-            None,
-            &mut prover_transcript,
-        )
-        .unwrap();
-
-        // Tamper with proof: swap v[0] and v[1]
-        let v1 = proof.v[1].clone();
-        proof.v[0].clone_from(&v1);
-
-        let mut verifier_transcript = Blake2bTranscript::new(b"test-tamper");
-        let result = <TestScheme as CommitmentScheme>::verify(
-            &commitment,
-            &point,
-            eval,
-            &proof,
-            &vk,
-            &mut verifier_transcript,
-        );
-        assert!(result.is_err(), "tampered proof should be rejected");
-    }
-
-    #[test]
-    fn combine_is_homomorphic() {
-        let ell = 3;
-        let n = 1 << ell;
-        let mut rng = ChaCha20Rng::seed_from_u64(300);
-        let (pk, _vk) = test_setup(n);
-
-        let poly_a = Polynomial::<Fr>::random(ell, &mut rng);
-        let poly_b = Polynomial::<Fr>::random(ell, &mut rng);
-
-        let (ca, ()) = TestScheme::commit(poly_a.evaluations(), &pk).unwrap();
-        let (cb, ()) = TestScheme::commit(poly_b.evaluations(), &pk).unwrap();
-
-        let sum_evals: Vec<Fr> = poly_a
-            .evaluations()
-            .iter()
-            .zip(poly_b.evaluations().iter())
-            .map(|(a, b)| *a + *b)
-            .collect();
-        let (c_sum_direct, ()) = TestScheme::commit(&sum_evals, &pk).unwrap();
-
-        let c_sum_combined = TestScheme::combine(&[ca, cb], &[Fr::from_u64(1), Fr::from_u64(1)]);
-
-        assert_eq!(
-            c_sum_direct, c_sum_combined,
-            "combine([1,1]) must match commitment to sum"
-        );
-    }
-
-    #[test]
-    fn combine_with_scalars() {
-        let ell = 3;
-        let n = 1 << ell;
-        let mut rng = ChaCha20Rng::seed_from_u64(400);
-        let (pk, _vk) = test_setup(n);
-
-        let poly_a = Polynomial::<Fr>::random(ell, &mut rng);
-        let poly_b = Polynomial::<Fr>::random(ell, &mut rng);
-        let s_a = Fr::random(&mut rng);
-        let s_b = Fr::random(&mut rng);
-
-        let (ca, ()) = TestScheme::commit(poly_a.evaluations(), &pk).unwrap();
-        let (cb, ()) = TestScheme::commit(poly_b.evaluations(), &pk).unwrap();
-
-        let combined_evals: Vec<Fr> = poly_a
-            .evaluations()
-            .iter()
-            .zip(poly_b.evaluations().iter())
-            .map(|(a, b)| s_a * *a + s_b * *b)
-            .collect();
-        let (c_direct, ()) = TestScheme::commit(&combined_evals, &pk).unwrap();
-
-        let c_combined = TestScheme::combine(&[ca, cb], &[s_a, s_b]);
-
-        assert_eq!(c_direct, c_combined);
-    }
-
-    #[test]
-    fn open_verify_with_random_points() {
-        let mut rng = ChaCha20Rng::seed_from_u64(0xcafe);
-
-        for _ in 0..5 {
-            let ell = 4;
-            let n = 1 << ell;
-            let (pk, vk) = test_setup(n);
-
-            let poly = Polynomial::<Fr>::random(ell, &mut rng);
-            let point: Vec<Fr> = (0..ell).map(|_| Fr::random(&mut rng)).collect();
-            let eval = poly.evaluate(&point);
-
-            let (commitment, ()) = TestScheme::commit(poly.evaluations(), &pk).unwrap();
-
-            let mut pt = Blake2bTranscript::new(b"rand-test");
-            let proof =
-                <TestScheme as CommitmentScheme>::open(&poly, &point, eval, &pk, None, &mut pt)
-                    .unwrap();
-
-            let mut vt = Blake2bTranscript::new(b"rand-test");
-            <TestScheme as CommitmentScheme>::verify(
-                &commitment,
-                &point,
-                eval,
-                &proof,
-                &vk,
-                &mut vt,
-            )
-            .expect("random instance should verify");
-        }
-    }
-
-    #[test]
-    fn extract_vc_setup_produces_valid_pedersen() {
-        use jolt_crypto::{Pedersen, VectorCommitment};
-
-        let n = 1 << 4;
-        let (pk, _vk) = test_setup(n);
-
-        let capacity = 5;
-        let vc_setup = PedersenSetup::<jolt_crypto::Bn254G1>::derive(&pk, capacity);
-
-        assert_eq!(
-            <Pedersen<jolt_crypto::Bn254G1> as VectorCommitment>::capacity(&vc_setup),
-            capacity,
-        );
-
-        // Commit and verify a small vector.
-        let values = vec![Fr::one(), Fr::from_u64(2), Fr::from_u64(3)];
-        let blinding = Fr::from_u64(42);
-        let commitment = <Pedersen<jolt_crypto::Bn254G1> as VectorCommitment>::commit(
-            &vc_setup, &values, &blinding,
-        );
-        assert!(
-            <Pedersen<jolt_crypto::Bn254G1> as VectorCommitment>::verify(
-                &vc_setup,
-                &commitment,
-                &values,
-                &blinding,
-            )
-        );
-    }
-
-    #[test]
-    fn trivial_polynomial() {
-        // 1-variable polynomial: [a, b]
-        let ell = 1;
-        let n = 1 << ell;
-        let mut rng = ChaCha20Rng::seed_from_u64(777);
-        let (pk, vk) = test_setup(n);
-
-        let poly = Polynomial::<Fr>::random(ell, &mut rng);
-        let point: Vec<Fr> = (0..ell).map(|_| Fr::random(&mut rng)).collect();
-        let eval = poly.evaluate(&point);
-
-        let (commitment, ()) = TestScheme::commit(poly.evaluations(), &pk).unwrap();
-
-        let mut pt = Blake2bTranscript::new(b"trivial");
-        let proof = <TestScheme as CommitmentScheme>::open(&poly, &point, eval, &pk, None, &mut pt)
-            .unwrap();
-
-        let mut vt = Blake2bTranscript::new(b"trivial");
-        <TestScheme as CommitmentScheme>::verify(&commitment, &point, eval, &proof, &vk, &mut vt)
-            .expect("trivial polynomial should verify");
-    }
-}
