+++ crates/jolt-openings/Cargo.toml
@@ -19,7 +19,6 @@ tracing.workspace = true
 
 [dev-dependencies]
 jolt-dory.workspace = true
-jolt-hyperkzg.workspace = true
 jolt-crypto = { workspace = true, features = ["bn254"] }
 rand.workspace = true
 rand_chacha.workspace = true
