--- crates/jolt-prover-legacy/src/poly/commitment/hyperkzg.rs
@@ -1,767 +1,0 @@
-//! This is a port of https://github.com/microsoft/Nova/blob/main/src/provider/hyperkzg.rs
-//! and such code is Copyright (c) Microsoft Corporation.
-//!
-//! This module implements `HyperKZG`, a KZG-based polynomial commitment for multilinear polynomials
-//! HyperKZG is based on the transformation from univariate PCS to multilinear PCS in the Gemini paper (section 2.4.2 in <https://eprint.iacr.org/2022/420.pdf>).
-//! However, there are some key differences:
-//! (1) HyperKZG works with multilinear polynomials represented in evaluation form (rather than in coefficient form in Gemini's transformation).
-//! This means that Spartan's polynomial IOP can use commit to its polynomials as-is without incurring any interpolations or FFTs.
-//! (2) HyperKZG is specialized to use KZG as the univariate commitment scheme, so it includes several optimizations (both during the transformation of multilinear-to-univariate claims
-//! and within the KZG commitment scheme implementation itself).
-use super::{
-    commitment_scheme::CommitmentScheme,
-    kzg::{KZGProverKey, KZGVerifierKey, UnivariateKZG},
-};
-use crate::field::JoltField;
-use crate::poly::multilinear_polynomial::{MultilinearPolynomial, PolynomialEvaluation};
-use crate::poly::rlc_polynomial::RLCPolynomial;
-use crate::zkvm::witness::CommittedPolynomial;
-use crate::{
-    msm::VariableBaseMSM,
-    poly::{commitment::kzg::SRS, dense_mlpoly::DensePolynomial, unipoly::UniPoly},
-    transcripts::Transcript,
-    utils::{errors::ProofVerifyError, small_scalar::SmallScalar},
-};
-use ark_ec::{pairing::Pairing, AffineRepr, CurveGroup};
-use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, Compress, Validate};
-use ark_std::{One, Zero};
-use rand::rngs::OsRng;
-use rand_core::{CryptoRng, RngCore};
-use rayon::iter::{
-    IndexedParallelIterator, IntoParallelIterator, IntoParallelRefIterator,
-    IntoParallelRefMutIterator, ParallelIterator,
-};
-use std::borrow::Borrow;
-use std::{io::Cursor, marker::PhantomData, sync::Arc};
-
-pub struct HyperKZGSRS<P: Pairing>(Arc<SRS<P>>);
-
-impl<P: Pairing> HyperKZGSRS<P> {
-    pub fn setup<R: RngCore + CryptoRng>(rng: &mut R, max_degree: usize) -> Self
-    where
-        P::ScalarField: JoltField,
-    {
-        Self(Arc::new(SRS::setup(rng, max_degree, 2)))
-    }
-
-    pub fn trim(self, max_degree: usize) -> (HyperKZGProverKey<P>, HyperKZGVerifierKey<P>) {
-        let (kzg_pk, kzg_vk) = SRS::trim(self.0, max_degree);
-        (HyperKZGProverKey { kzg_pk }, HyperKZGVerifierKey { kzg_vk })
-    }
-}
-
-#[derive(Clone, Debug, CanonicalSerialize, CanonicalDeserialize)]
-pub struct HyperKZGProverKey<P: Pairing> {
-    pub kzg_pk: KZGProverKey<P>,
-}
-
-#[derive(Copy, Clone, Debug, CanonicalSerialize, CanonicalDeserialize)]
-pub struct HyperKZGVerifierKey<P: Pairing> {
-    pub kzg_vk: KZGVerifierKey<P>,
-}
-
-#[derive(Debug, Clone, PartialEq, CanonicalSerialize, CanonicalDeserialize)]
-pub struct HyperKZGCommitment<P: Pairing>(pub P::G1Affine);
-
-impl<P: Pairing> Default for HyperKZGCommitment<P> {
-    fn default() -> Self {
-        Self(P::G1Affine::zero())
-    }
-}
-
-#[derive(Clone, CanonicalSerialize, CanonicalDeserialize, Debug)]
-pub struct HyperKZGProof<P: Pairing> {
-    pub com: Vec<P::G1Affine>,
-    pub w: Vec<P::G1Affine>,
-    pub v: Vec<Vec<P::ScalarField>>,
-}
-
-// On input f(x) and u compute the witness polynomial used to prove
-// that f(u) = v. The main part of this is to compute the
-// division (f(x) - f(u)) / (x - u), but we don't use a general
-// division algorithm, we make use of the fact that the division
-// never has a remainder, and that the denominator is always a linear
-// polynomial. The cost is (d-1) mults + (d-1) adds in P::ScalarField, where
-// d is the degree of f.
-//
-// We use the fact that if we compute the quotient of f(x)/(x-u),
-// there will be a remainder, but it'll be v = f(u).  Put another way
-// the quotient of f(x)/(x-u) and (f(x) - f(v))/(x-u) is the
-// same.  One advantage is that computing f(u) could be decoupled
-// from kzg_open, it could be done later or separate from computing W.
-fn kzg_batch_open_no_rem<P: Pairing>(
-    f: &MultilinearPolynomial<P::ScalarField>,
-    u: &[P::ScalarField],
-    pk: &HyperKZGProverKey<P>,
-) -> Vec<P::G1Affine>
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    let f: &DensePolynomial<P::ScalarField> = f.try_into().unwrap();
-    let h = u
-        .par_iter()
-        .map(|ui| {
-            let h = compute_witness_polynomial::<P>(&f.evals(), *ui);
-            MultilinearPolynomial::from(h)
-        })
-        .collect::<Vec<_>>();
-
-    UnivariateKZG::commit_batch(&pk.kzg_pk, &h).unwrap()
-}
-
-fn compute_witness_polynomial<P: Pairing>(
-    f: &[P::ScalarField],
-    u: P::ScalarField,
-) -> Vec<P::ScalarField>
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    let d = f.len();
-
-    // Compute h(x) = f(x)/(x - u)
-    let mut h = vec![P::ScalarField::zero(); d];
-    for i in (1..d).rev() {
-        h[i - 1] = f[i] + h[i] * u;
-    }
-
-    h
-}
-
-fn absorb_hyperkzg_witness_commitments<P: Pairing, ProofTranscript: Transcript>(
-    transcript: &mut ProofTranscript,
-    witness_commitments: &[P::G1Affine],
-) -> P::ScalarField
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    transcript.append_commitments_serializable(b"hyperkzg_witness", witness_commitments);
-    transcript.challenge_scalar()
-}
-
-fn validate_hyperkzg_g1_point<P: Pairing>(point: &P::G1Affine) -> Result<(), ProofVerifyError>
-where
-    P::G1Affine: CanonicalSerialize + CanonicalDeserialize,
-{
-    if point.is_zero() {
-        return Err(ProofVerifyError::InternalError);
-    }
-
-    let mut encoded = Vec::new();
-    point
-        .serialize_with_mode(&mut encoded, Compress::No)
-        .map_err(|_| ProofVerifyError::InternalError)?;
-    let cursor = Cursor::new(encoded);
-    P::G1Affine::deserialize_with_mode(cursor, Compress::No, Validate::Yes)
-        .map_err(|_| ProofVerifyError::InternalError)?;
-
-    Ok(())
-}
-
-fn kzg_open_batch<P: Pairing, ProofTranscript: Transcript>(
-    f: &[MultilinearPolynomial<P::ScalarField>],
-    u: &[P::ScalarField],
-    pk: &HyperKZGProverKey<P>,
-    transcript: &mut ProofTranscript,
-) -> (Vec<P::G1Affine>, Vec<Vec<P::ScalarField>>)
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    let k = f.len();
-    let t = u.len();
-
-    // The verifier needs f_i(u_j), so we compute them here
-    // (V will compute B(u_j) itself)
-    let mut v = vec![vec!(P::ScalarField::zero(); k); t];
-    v.par_iter_mut().enumerate().for_each(|(i, v_i)| {
-        // for each point u
-        v_i.par_iter_mut().zip_eq(f).for_each(|(v_ij, f)| {
-            // for each poly f
-            *v_ij = UniPoly::eval_as_univariate(f, &u[i]);
-        });
-    });
-
-    // TODO(moodlezoup): Avoid cloned()
-    let scalars: Vec<P::ScalarField> = v.iter().flatten().cloned().collect();
-    transcript.append_scalars(b"hyperkzg_evals", &scalars);
-    let q_powers: Vec<P::ScalarField> = transcript.challenge_scalar_powers(f.len());
-    let f_arc: Vec<Arc<MultilinearPolynomial<P::ScalarField>>> =
-        f.iter().map(|poly| Arc::new(poly.clone())).collect();
-
-    // HyperKZG is not currently used in Jolt, but we handle both dense-only and mixed cases
-    let has_one_hot = f_arc
-        .iter()
-        .any(|poly| matches!(poly.as_ref(), MultilinearPolynomial::OneHot(_)));
-
-    let B = if has_one_hot {
-        // Use RLCPolynomial::linear_combination for mixed dense + one-hot polynomials
-        let dummy_poly_ids = vec![CommittedPolynomial::RdInc; f_arc.len()];
-        let rlc_result = RLCPolynomial::linear_combination(dummy_poly_ids, f_arc, &q_powers, None);
-        MultilinearPolynomial::RLC(rlc_result)
-    } else {
-        let poly_refs: Vec<&MultilinearPolynomial<P::ScalarField>> =
-            f_arc.iter().map(|arc| arc.as_ref()).collect();
-        let dense_result = DensePolynomial::linear_combination(&poly_refs, &q_powers);
-        MultilinearPolynomial::from(dense_result.Z)
-    };
-
-    // Now open B at u0, ..., u_{t-1}
-    let w = kzg_batch_open_no_rem(&B, u, pk);
-
-    // The prover and verifier must absorb the witness commitments identically
-    // so they derive the same Fiat-Shamir challenge.
-    let _d_0: P::ScalarField =
-        absorb_hyperkzg_witness_commitments::<P, ProofTranscript>(transcript, &w);
-
-    (w, v)
-}
-
-// vk is hashed in transcript already, so we do not add it here
-fn kzg_verify_batch<P: Pairing, ProofTranscript: Transcript>(
-    vk: &HyperKZGVerifierKey<P>,
-    C: &[P::G1Affine],
-    W: &[P::G1Affine],
-    u: &[P::ScalarField],
-    v: &[Vec<P::ScalarField>],
-    transcript: &mut ProofTranscript,
-) -> bool
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    let k = C.len();
-    let t = u.len();
-
-    let scalars: Vec<P::ScalarField> = v.iter().flatten().cloned().collect();
-    transcript.append_scalars(b"hyperkzg_evals", &scalars);
-    let q_powers: Vec<P::ScalarField> = transcript.challenge_scalar_powers(k);
-
-    let d_0: P::ScalarField =
-        absorb_hyperkzg_witness_commitments::<P, ProofTranscript>(transcript, W);
-    let d_1 = d_0 * d_0;
-
-    assert_eq!(t, 3);
-    assert_eq!(W.len(), 3);
-    // We write a special case for t=3, since this what is required for
-    // hyperkzg. Following the paper directly, we must compute:
-    // let L0 = C_B - vk.G * B_u[0] + W[0] * u[0];
-    // let L1 = C_B - vk.G * B_u[1] + W[1] * u[1];
-    // let L2 = C_B - vk.G * B_u[2] + W[2] * u[2];
-    // let R0 = -W[0];
-    // let R1 = -W[1];
-    // let R2 = -W[2];
-    // let L = L0 + L1*d_0 + L2*d_1;
-    // let R = R0 + R1*d_0 + R2*d_1;
-    //
-    // We group terms to reduce the number of scalar mults (to seven):
-    // In Rust, we could use MSMs for these, and speed up verification.
-    //
-    // Note, that while computing L, the intermediate computation of C_B together with computing
-    // L0, L1, L2 can be replaced by single MSM of C with the powers of q multiplied by (1 + d_0 + d_1)
-    // with additionally concatenated inputs for scalars/bases.
-
-    let q_power_multiplier: P::ScalarField = P::ScalarField::one() + d_0 + d_1;
-
-    let q_powers_multiplied: Vec<P::ScalarField> = q_powers
-        .par_iter()
-        .map(|q_power| *q_power * q_power_multiplier)
-        .collect();
-
-    // Compute the batched openings
-    // compute B(u_i) = v[i][0] + q*v[i][1] + ... + q^(t-1) * v[i][t-1]
-    let B_u = v
-        .into_par_iter()
-        .map(|v_i| {
-            v_i.into_par_iter()
-                .zip(q_powers.par_iter())
-                .map(|(a, b)| *a * *b)
-                .sum()
-        })
-        .collect::<Vec<P::ScalarField>>();
-
-    let L = <P::G1 as VariableBaseMSM>::msm_field_elements(
-        &[&C[..k], &[W[0], W[1], W[2], vk.kzg_vk.g1]].concat(),
-        &[
-            &q_powers_multiplied[..k],
-            &[
-                u[0],
-                (u[1] * d_0),
-                (u[2] * d_1),
-                -(B_u[0] + d_0 * B_u[1] + d_1 * B_u[2]),
-            ],
-        ]
-        .concat(),
-    )
-    .unwrap();
-
-    let R = W[0] + W[1] * d_0 + W[2] * d_1;
-
-    // Check that e(L, vk.H) == e(R, vk.tau_H)
-    P::multi_pairing([L, -R], [vk.kzg_vk.g2, vk.kzg_vk.beta_g2]).is_zero()
-}
-
-#[derive(Clone)]
-pub struct HyperKZG<P: Pairing> {
-    _phantom: PhantomData<P>,
-}
-
-impl<P: Pairing> HyperKZG<P>
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    pub fn protocol_name() -> &'static [u8] {
-        b"HyperKZG"
-    }
-
-    pub fn commit(
-        pp: &HyperKZGProverKey<P>,
-        poly: &MultilinearPolynomial<P::ScalarField>,
-    ) -> Result<HyperKZGCommitment<P>, ProofVerifyError> {
-        if pp.kzg_pk.g1_powers().len() < poly.len() {
-            return Err(ProofVerifyError::KeyLengthError(
-                pp.kzg_pk.g1_powers().len(),
-                poly.len(),
-            ));
-        }
-        Ok(HyperKZGCommitment(UnivariateKZG::commit_as_univariate(
-            &pp.kzg_pk, poly,
-        )?))
-    }
-
-    #[tracing::instrument(skip_all, name = "HyperKZG::open")]
-    pub fn open<ProofTranscript: Transcript>(
-        pk: &HyperKZGProverKey<P>,
-        poly: &MultilinearPolynomial<P::ScalarField>,
-        point: &[<P::ScalarField as JoltField>::Challenge],
-        _eval: &P::ScalarField,
-        transcript: &mut ProofTranscript,
-    ) -> Result<HyperKZGProof<P>, ProofVerifyError> {
-        let ell = point.len();
-        let n = poly.len();
-        assert_eq!(n, 1 << ell); // Below we assume that n is a power of two
-
-        // Phase 1  -- create commitments com_1, ..., com_\ell
-        // We do not compute final Pi (and its commitment) as it is constant and equals to 'eval'
-        // also known to verifier, so can be derived on its side as well
-        let mut polys: Vec<MultilinearPolynomial<P::ScalarField>> = Vec::new();
-        polys.push(poly.clone());
-        for i in 0..ell - 1 {
-            let previous_poly: &DensePolynomial<P::ScalarField> = (&polys[i]).try_into().unwrap();
-            let Pi_len = previous_poly.len() / 2;
-            let mut Pi = vec![P::ScalarField::zero(); Pi_len];
-            Pi.par_iter_mut().enumerate().for_each(|(j, Pi_j)| {
-                *Pi_j = point[ell - i - 1] * (previous_poly[2 * j + 1] - previous_poly[2 * j])
-                    + previous_poly[2 * j];
-            });
-
-            polys.push(MultilinearPolynomial::from(Pi));
-        }
-
-        assert_eq!(polys.len(), ell);
-        assert_eq!(polys[ell - 1].len(), 2);
-
-        // We do not need to commit to the first polynomial as it is already committed.
-        let com: Vec<P::G1Affine> = UnivariateKZG::commit_variable_batch(&pk.kzg_pk, &polys[1..])?;
-
-        // Phase 2
-        // We do not need to add x to the transcript, because in our context x was obtained from the transcript.
-        // We also do not need to absorb `C` and `eval` as they are already absorbed by the transcript by the caller
-        transcript.append_commitments_serializable(b"hyperkzg_com", &com);
-        let r: <P as Pairing>::ScalarField = transcript.challenge_scalar();
-        let u = vec![r, -r, r * r];
-
-        // Phase 3 -- create response
-        let (w, v) = kzg_open_batch(&polys, &u, pk, transcript);
-
-        Ok(HyperKZGProof { com, w, v })
-    }
-
-    /// A method to verify purported evaluations of a batch of polynomials
-    pub fn verify<ProofTranscript: Transcript>(
-        vk: &HyperKZGVerifierKey<P>,
-        C: &HyperKZGCommitment<P>,
-        point: &[<P::ScalarField as JoltField>::Challenge],
-        P_of_x: &P::ScalarField,
-        pi: &HyperKZGProof<P>,
-        transcript: &mut ProofTranscript,
-    ) -> Result<(), ProofVerifyError> {
-        let y = P_of_x;
-
-        let ell = point.len();
-
-        validate_hyperkzg_g1_point::<P>(&C.0)?;
-        for commitment in &pi.com {
-            validate_hyperkzg_g1_point::<P>(commitment)?;
-        }
-        for witness in &pi.w {
-            validate_hyperkzg_g1_point::<P>(witness)?;
-        }
-
-        let mut com = pi.com.clone();
-
-        // we do not need to add x to the transcript, because in our context x was
-        // obtained from the transcript
-        transcript.append_commitments_serializable(b"hyperkzg_com", &com);
-        let r: <P as Pairing>::ScalarField = transcript.challenge_scalar();
-
-        if r == P::ScalarField::zero() {
-            return Err(ProofVerifyError::InternalError);
-        }
-        com.insert(0, C.0); // set com_0 = C, shifts other commitments to the right
-
-        let u = vec![r, -r, r * r];
-
-        // Setup vectors (Y, ypos, yneg) from pi.v
-        let v = &pi.v;
-        if v.len() != 3 {
-            return Err(ProofVerifyError::InternalError);
-        }
-        if v[0].len() != ell || v[1].len() != ell || v[2].len() != ell {
-            return Err(ProofVerifyError::InternalError);
-        }
-        let ypos = &v[0];
-        let yneg = &v[1];
-        let mut Y = v[2].to_vec();
-        Y.push(*y);
-
-        // Check consistency of (Y, ypos, yneg)
-        let two = P::ScalarField::from(2u64);
-        for i in 0..ell {
-            if two * r * Y[i + 1]
-                != r * (P::ScalarField::one() - point[ell - i - 1]) * (ypos[i] + yneg[i])
-                    + point[ell - i - 1] * (ypos[i] - yneg[i])
-            {
-                return Err(ProofVerifyError::InternalError);
-            }
-            // Note that we don't make any checks about Y[0] here, but our batching
-            // check below requires it
-        }
-
-        // Check commitments to (Y, ypos, yneg) are valid
-        if !kzg_verify_batch(vk, &com, &pi.w, &u, &pi.v, transcript) {
-            return Err(ProofVerifyError::InternalError);
-        }
-
-        Ok(())
-    }
-}
-
-impl<P: Pairing> CommitmentScheme for HyperKZG<P>
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    type Field = P::ScalarField;
-    type ProverSetup = HyperKZGProverKey<P>;
-    type VerifierSetup = HyperKZGVerifierKey<P>;
-
-    type Commitment = HyperKZGCommitment<P>;
-    type Proof = HyperKZGProof<P>;
-    type BatchedProof = HyperKZGProof<P>;
-    type OpeningProofHint = ();
-
-    fn setup_prover(max_num_vars: usize) -> Self::ProverSetup {
-        let mut rng = OsRng;
-        HyperKZGSRS(Arc::new(SRS::setup(&mut rng, 1 << max_num_vars, 2)))
-            .trim(1 << max_num_vars)
-            .0
-    }
-
-    fn setup_verifier(setup: &Self::ProverSetup) -> Self::VerifierSetup {
-        HyperKZGVerifierKey {
-            kzg_vk: KZGVerifierKey::from(&setup.kzg_pk),
-        }
-    }
-
-    #[tracing::instrument(skip_all, name = "HyperKZG::commit")]
-    fn commit(
-        poly: &MultilinearPolynomial<Self::Field>,
-        setup: &Self::ProverSetup,
-    ) -> (Self::Commitment, Self::OpeningProofHint) {
-        assert!(
-            setup.kzg_pk.g1_powers().len() >= poly.len(),
-            "COMMIT KEY LENGTH ERROR {}, {}",
-            setup.kzg_pk.g1_powers().len(),
-            poly.len()
-        );
-        let commitment =
-            HyperKZGCommitment(UnivariateKZG::commit_as_univariate(&setup.kzg_pk, poly).unwrap());
-        (commitment, ())
-    }
-
-    #[tracing::instrument(skip_all, name = "HyperKZG::batch_commit")]
-    fn batch_commit<U>(
-        polys: &[U],
-        gens: &Self::ProverSetup,
-    ) -> Vec<(Self::Commitment, Self::OpeningProofHint)>
-    where
-        U: Borrow<MultilinearPolynomial<Self::Field>> + Sync,
-    {
-        UnivariateKZG::commit_batch(&gens.kzg_pk, polys)
-            .unwrap()
-            .into_par_iter()
-            .map(|c| (HyperKZGCommitment(c), ()))
-            .collect()
-    }
-
-    fn combine_commitments<C: Borrow<Self::Commitment>>(
-        commitments: &[C],
-        coeffs: &[Self::Field],
-    ) -> Self::Commitment {
-        let combined_commitment: P::G1 = commitments
-            .iter()
-            .zip(coeffs.iter())
-            .map(|(commitment, coeff)| commitment.borrow().0 * coeff)
-            .sum();
-        HyperKZGCommitment(combined_commitment.into_affine())
-    }
-
-    fn prove<ProofTranscript: Transcript>(
-        setup: &Self::ProverSetup,
-        poly: &MultilinearPolynomial<Self::Field>,
-        opening_point: &[<Self::Field as JoltField>::Challenge], // point at which the polynomial is evaluated
-        _hint: Option<Self::OpeningProofHint>,
-        transcript: &mut ProofTranscript,
-    ) -> (Self::Proof, Option<Self::Field>) {
-        let eval = poly.evaluate(opening_point);
-        let proof = HyperKZG::<P>::open(setup, poly, opening_point, &eval, transcript).unwrap();
-        (proof, None) // HyperKZG doesn't have ZK blinding
-    }
-
-    fn verify<ProofTranscript: Transcript>(
-        proof: &Self::Proof,
-        setup: &Self::VerifierSetup,
-        transcript: &mut ProofTranscript,
-        opening_point: &[<Self::Field as JoltField>::Challenge], // point at which the polynomial is evaluated
-        opening: &Self::Field,                                   // evaluation \widetilde{Z}(r)
-        commitment: &Self::Commitment,
-    ) -> Result<(), ProofVerifyError> {
-        HyperKZG::<P>::verify(setup, commitment, opening_point, opening, proof, transcript)
-    }
-
-    fn protocol_name() -> &'static [u8] {
-        b"hyperkzg"
-    }
-}
-
-impl<P: Pairing> super::commitment_scheme::StreamingCommitmentScheme for HyperKZG<P>
-where
-    <P as Pairing>::ScalarField: JoltField,
-{
-    type ChunkState = ();
-
-    fn process_chunk<T: SmallScalar>(_setup: &Self::ProverSetup, _chunk: &[T]) -> Self::ChunkState {
-    }
-
-    fn process_chunk_onehot(
-        _setup: &Self::ProverSetup,
-        _onehot_k: usize,
-        _chunk: &[Option<usize>],
-    ) -> Self::ChunkState {
-    }
-
-    fn aggregate_chunks(
-        _setup: &Self::ProverSetup,
-        _onehot_k: Option<usize>,
-        _tier1_commitments: &[Self::ChunkState],
-    ) -> (Self::Commitment, Self::OpeningProofHint) {
-        unimplemented!("HyperKZG does not support streaming commitment")
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::transcripts::{Blake2bTranscript, Transcript};
-    use ark_bn254::{Bn254, Fq};
-    use ark_ec::AffineRepr;
-    use ark_std::UniformRand;
-    use ark_std::Zero;
-    use rand::Rng;
-    use rand_chacha::ChaCha20Rng;
-    use rand_core::SeedableRng;
-
-    //#[test]
-    //fn test_hyperkzg_eval() {
-    //    // Test with poly(X1, X2) = 1 + X1 + X2 + X1*X2
-    //    let mut rng = rand_chacha::ChaCha20Rng::seed_from_u64(0);
-    //    let srs = HyperKZGSRS::setup(&mut rng, 3);
-    //    let (pk, vk): (HyperKZGProverKey<Bn254>, HyperKZGVerifierKey<Bn254>) = srs.trim(3);
-    //
-    //    // poly is in eval. representation; evaluated at [(0,0), (0,1), (1,0), (1,1)]
-    //    let poly =
-    //        MultilinearPolynomial::from(vec![Fr::from(1), Fr::from(2), Fr::from(2), Fr::from(4)]);
-    //
-    //    let C = HyperKZG::commit(&pk, &poly).unwrap();
-    //
-    //    let test_inner =
-    //        |point: Vec<MontU128Challenge<Fr>>, eval: Fr| -> Result<(), ProofVerifyError> {
-    //            let mut tr = Blake2bTranscript::new(b"TestEval");
-    //            let proof = HyperKZG::open(&pk, &poly, &point, &eval, &mut tr).unwrap();
-    //            let mut tr = Blake2bTranscript::new(b"TestEval");
-    //            HyperKZG::verify(&vk, &C, &point, &eval, &proof, &mut tr)
-    //        };
-    //
-    //    // Call the prover with a (point, eval) pair.
-    //    // The prover does not recompute so it may produce a proof, but it should not verify
-    //    let point = vec![Fr::from(0), Fr::from(0)];
-    //    let eval = Fr::from(1);
-    //    assert!(test_inner(point, eval).is_ok());
-    //
-    //    let point = vec![Fr::from(0), Fr::from(1)];
-    //    let eval = Fr::from(2);
-    //    assert!(test_inner(point, eval).is_ok());
-    //
-    //    let point = vec![Fr::from(1), Fr::from(1)];
-    //    let eval = Fr::from(4);
-    //    assert!(test_inner(point, eval).is_ok());
-    //
-    //    let point = vec![Fr::from(0), Fr::from(2)];
-    //    let eval = Fr::from(3);
-    //    assert!(test_inner(point, eval).is_ok());
-    //
-    //    let point = vec![Fr::from(2), Fr::from(2)];
-    //    let eval = Fr::from(9);
-    //    assert!(test_inner(point, eval).is_ok());
-    //
-    //    // Try a couple incorrect evaluations and expect failure
-    //    let point = vec![Fr::from(2), Fr::from(2)];
-    //    let eval = Fr::from(50);
-    //    assert!(test_inner(point, eval).is_err());
-    //
-    //    let point = vec![Fr::from(0), Fr::from(2)];
-    //    let eval = Fr::from(4);
-    //    assert!(test_inner(point, eval).is_err());
-    //}
-
-    // THIS test does not make sense for MontU128Challenge
-    //#[test]
-    //fn test_hyperkzg_small() {
-    //    let mut rng = rand_chacha::ChaCha20Rng::seed_from_u64(0);
-    //
-    //    // poly = [1, 2, 1, 4]
-    //    let poly =
-    //        MultilinearPolynomial::from(vec![Fr::from(1), Fr::from(2), Fr::from(1), Fr::from(4)]);
-    //
-    //    // point = [4,3]
-    //    let point = vec![Fr::from(4), Fr::from(3)];
-    //
-    //    // eval = 28
-    //    let eval = Fr::from(28);
-    //
-    //    let srs = HyperKZGSRS::setup(&mut rng, 3);
-    //    let (pk, vk): (HyperKZGProverKey<Bn254>, HyperKZGVerifierKey<Bn254>) = srs.trim(3);
-    //
-    //    // make a commitment
-    //    let C = HyperKZG::commit(&pk, &poly).unwrap();
-    //
-    //    // prove an evaluation
-    //    let mut tr = Blake2bTranscript::new(b"TestEval");
-    //    let proof = HyperKZG::open(&pk, &poly, &point, &eval, &mut tr).unwrap();
-    //    let post_c_p = tr.challenge_scalar::<Fr>();
-    //
-    //    // verify the evaluation
-    //    let mut verifier_transcript = Blake2bTranscript::new(b"TestEval");
-    //    assert!(
-    //        HyperKZG::verify(&vk, &C, &point, &eval, &proof, &mut verifier_transcript,).is_ok()
-    //    );
-    //    let post_c_v = verifier_transcript.challenge_scalar::<Fr>();
-    //
-    //    // check if the prover transcript and verifier transcript are kept in the same state
-    //    assert_eq!(post_c_p, post_c_v);
-    //
-    //    let mut proof_bytes = Vec::new();
-    //    proof.serialize_compressed(&mut proof_bytes).unwrap();
-    //    assert_eq!(proof_bytes.len(), 368);
-    //
-    //    // Change the proof and expect verification to fail
-    //    let mut bad_proof = proof.clone();
-    //    let v1 = bad_proof.v[1].clone();
-    //    bad_proof.v[0].clone_from(&v1);
-    //    let mut verifier_transcript2 = Blake2bTranscript::new(b"TestEval");
-    //    assert!(HyperKZG::verify(
-    //        &vk,
-    //        &C,
-    //        &point,
-    //        &eval,
-    //        &bad_proof,
-    //        &mut verifier_transcript2
-    //    )
-    //    .is_err());
-    //}
-    #[test]
-    fn test_hyperkzg_large() {
-        // test the hyperkzg prover and verifier with random instances (derived from a seed)
-        for ell in [8, 9, 10] {
-            let mut rng = ChaCha20Rng::seed_from_u64(ell as u64);
-
-            let n = 1 << ell; // n = 2^ell
-
-            let poly_raw = (0..n)
-                .map(|_| <Bn254 as Pairing>::ScalarField::rand(&mut rng))
-                .collect::<Vec<_>>();
-            let poly = MultilinearPolynomial::from(poly_raw.clone());
-            let point = (0..ell)
-                .map(|_| {
-                    <<Bn254 as Pairing>::ScalarField as JoltField>::Challenge::from(
-                        rng.gen::<u128>(),
-                    )
-                })
-                .collect::<Vec<_>>();
-            let eval = poly.evaluate(&point);
-
-            let srs = HyperKZGSRS::setup(&mut rng, n);
-            let (pk, vk): (HyperKZGProverKey<Bn254>, HyperKZGVerifierKey<Bn254>) = srs.trim(n);
-            assert_eq!(vk.kzg_vk.g1, pk.kzg_pk.g1_powers()[0]);
-            assert_eq!(vk.kzg_vk.alpha_g1, pk.kzg_pk.g1_powers()[1]);
-
-            // make a commitment
-            let C = HyperKZG::commit(&pk, &poly).unwrap();
-
-            // prove an evaluation
-            let mut prover_transcript = Blake2bTranscript::new(b"TestEval");
-            let proof: HyperKZGProof<Bn254> =
-                HyperKZG::open(&pk, &poly, &point, &eval, &mut prover_transcript).unwrap();
-
-            // verify the evaluation
-            let mut verifier_tr = Blake2bTranscript::new(b"TestEval");
-            assert!(HyperKZG::verify(&vk, &C, &point, &eval, &proof, &mut verifier_tr,).is_ok());
-            let post_c_p = prover_transcript.challenge_scalar::<ark_bn254::Fr>();
-            let post_c_v = verifier_tr.challenge_scalar::<ark_bn254::Fr>();
-            assert_eq!(post_c_p, post_c_v);
-
-            // Change the proof and expect verification to fail
-            let mut bad_proof = proof.clone();
-            let v1 = bad_proof.v[1].clone();
-            bad_proof.v[0].clone_from(&v1);
-            let mut verifier_tr2 = Blake2bTranscript::new(b"TestEval");
-            assert!(
-                HyperKZG::verify(&vk, &C, &point, &eval, &bad_proof, &mut verifier_tr2,).is_err()
-            );
-
-            let mut bad_identity_commitment = C.clone();
-            bad_identity_commitment.0 = <Bn254 as Pairing>::G1Affine::zero();
-            let mut verifier_tr3 = Blake2bTranscript::new(b"TestEval");
-            assert!(HyperKZG::verify(
-                &vk,
-                &bad_identity_commitment,
-                &point,
-                &eval,
-                &proof,
-                &mut verifier_tr3,
-            )
-            .is_err());
-
-            let mut bad_offcurve_proof = proof.clone();
-            bad_offcurve_proof.w[0] =
-                <Bn254 as Pairing>::G1Affine::new_unchecked(Fq::zero(), Fq::zero());
-            let mut verifier_tr4 = Blake2bTranscript::new(b"TestEval");
-            assert!(HyperKZG::verify(
-                &vk,
-                &C,
-                &point,
-                &eval,
-                &bad_offcurve_proof,
-                &mut verifier_tr4,
-            )
-            .is_err());
-        }
-    }
-}
