+++ crates/jolt-prover-legacy/src/poly/commitment/mod.rs
@@ -1,8 +1,6 @@
 pub mod commitment_scheme;
 pub mod dory;
-pub mod hyperkzg;
 pub mod hyrax;
-pub mod kzg;
 pub mod pedersen;
 
 #[cfg(test)]
